+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [feat]: `setComponentSelf()` method;
 - [feat]: `set()` method;
 - [feat]: `normalizeSelf()` method;
+- [feat]: update input options for `add()`, `cross()` methods;
 
 ## 0.10.0 (2022-11-28)